+++ deepinv/optim/data_fidelity.py
@@ -58,8 +58,8 @@ class DataFidelity(Potential):
         Computes the gradient :math:`\nabla_u\distance{u}{y}`, computed in :math:`u`.
 
         Note that this is the gradient of
-        :math:`\distancename` and not :math:`\datafidname`. This function direclty calls :meth:`deepinv.optim.Distance.grad` for the
-        speficic distance function :math:`\distancename`.
+        :math:`\distancename` and not :math:`\datafidname`. This function directly calls :meth:`deepinv.optim.Distance.grad` for the
+        specific distance function :math:`\distancename`.
 
         :param torch.Tensor u: Variable :math:`u` at which the gradient is computed.
         :param torch.Tensor y: Data :math:`y` of the same dimension as :math:`u`.
@@ -72,8 +72,8 @@ class DataFidelity(Potential):
         Computes the proximity operator :math:`\operatorname{prox}_{\gamma\distance{\cdot}{y}}(u)`, computed in :math:`u`.
 
         Note that this is the proximity operator of :math:`\distancename` and not :math:`\datafidname`.
-        This function direclty calls :meth:`deepinv.optim.Distance.prox` for the
-        speficic distance function :math:`\distancename`.
+        This function directly calls :meth:`deepinv.optim.Distance.prox` for the
+        specific distance function :math:`\distancename`.
 
         :param torch.Tensor u: Variable :math:`u` at which the gradient is computed.
         :param torch.Tensor y: Data :math:`y` of the same dimension as :math:`u`.
@@ -89,6 +89,124 @@ class DataFidelity(Potential):
         specific distance function :math:`\distancename`.
         """
         return self.d.prox_conjugate(u, y, *args, **kwargs)
+
+
+class StackedPhysicsDataFidelity(DataFidelity):
+    r"""
+    Stacked data fidelity term :math:`\datafid{x}{y} = \sum_i d_i(A_i(x),y_i})`.
+
+    Adapted to :class:`deepinv.physics.StackedPhysics` physics composed of multiple physics operators.
+
+    :param list[dinv.optim.DataFidelity] data_fidelity_list: list of data fidelity terms, one per physics operator.
+
+    |sep|
+
+    :Examples:
+
+        Define a stacked data fidelity term with two data fidelity terms :math:`\datafid{x}{y_1} + \datafid{x}{y_2}`:
+
+        >>> import torch
+        >>> import deepinv as dinv
+        >>> # define two observations, one with Gaussian noise and one with Poisson noise
+        >>> physics1 = dinv.physics.Denoising(dinv.physics.GaussianNoise(.1))
+        >>> physics2 = dinv.physics.Denoising(dinv.physics.PoissonNoise(.1))
+        >>> physics = dinv.physics.StackedLinearPhysics([physics1, physics2])
+        >>> fid1 = dinv.optim.L2()
+        >>> fid2 = dinv.optim.PoissonLikelihood()
+        >>> data_fidelity = dinv.optim.StackedPhysicsDataFidelity([fid1, fid2])
+        >>> x = torch.ones(1, 1, 3, 3) # image
+        >>> y = physics(x) # noisy measurements
+        >>> d = data_fidelity(x, y, physics)
+
+    """
+
+    def __init__(self, data_fidelity_list):
+        super(StackedPhysicsDataFidelity, self).__init__()
+        self.data_fidelity_list = data_fidelity_list
+
+    def fn(self, x, y, physics, *args, **kwargs):
+        r"""
+        Computes the data fidelity term :math:`\datafid{x}{y} = \sum_i d_i(A_i(x),y_i})`.
+
+        :param torch.Tensor x: Variable :math:`x` at which the data fidelity is computed.
+        :param deepinv.utils.TensorList y: Stacked measurements :math:`y`.
+        :param deepinv.physics.StackedPhysics physics: physics model.
+        :return: (torch.Tensor) data fidelity :math:`\datafid{x}{y}`.
+        """
+        out = 0
+        for i, data_fidelity in enumerate(self.data_fidelity_list):
+            out += data_fidelity.fn(x, y[i], physics[i], *args, **kwargs)
+        return out
+
+    def grad(self, x, y, physics, *args, **kwargs):
+        r"""
+        Calculates the gradient of the data fidelity term :math:`\datafidname` at :math:`x`.
+
+        The gradient is computed using the chain rule:
+
+        .. math::
+
+            \nabla_x \distance{\forw{x}}{y} = \sum_i \left. \frac{\partial A_i}{\partial x} \right|_x^\top \nabla_u \distance{u}{y_i},
+
+        where :math:`\left. \frac{\partial A_i}{\partial x} \right|_x` is the Jacobian of :math:`A_i` at :math:`x`,
+        and :math:`\nabla_u \distance{u}{y_i}` is computed using ``grad_d`` with :math:`u = \forw{x}`.
+        The multiplication is computed using the ``A_vjp`` method of each physics.
+
+        :param torch.Tensor x: Variable :math:`x` at which the gradient is computed.
+        :param dinv.utils.TensorList y: Stacked measurements :math:`y`.
+        :param deepinv.physics.StackedPhysics physics: Stacked physics model.
+        :return: (torch.Tensor) gradient :math:`\nabla_x \datafid{x}{y}`, computed in :math:`x`.
+        """
+        out = 0
+        for i, data_fidelity in enumerate(self.data_fidelity_list):
+            out += data_fidelity.grad(x, y[i], physics[i], *args, **kwargs)
+        return out
+
+    def grad_d(self, u, y, *args, **kwargs):
+        r"""
+        Computes the gradient :math:`\nabla_u\distance{u}{y}`, computed in :math:`u`.
+
+        Note that this is the gradient of
+        :math:`\distancename` and not :math:`\datafidname`. This function directly calls :meth:`deepinv.optim.Distance.grad` for the
+        specific distance function :math:`\distancename_i`.
+
+        :param torch.Tensor u: Variable :math:`u` at which the gradient is computed.
+        :param torch.Tensor y: Data :math:`y` of the same dimension as :math:`u`.
+        :return: (torch.Tensor) gradient of :math:`d` in :math:`u`, i.e. :math:`\nabla_u\distance{u}{y}`.
+        """
+        out = 0
+        for i, data_fidelity in enumerate(self.data_fidelity_list):
+            out += data_fidelity.grad_d(u, y[i], *args, **kwargs)
+        return out
+
+    def prox_d(self, u, y, *args, **kwargs):
+        r"""
+        Computes the proximity operator :math:`\operatorname{prox}_{\gamma\distance{\cdot}{y}}(u)`, computed in :math:`u`.
+
+        Note that this is the proximity operator of :math:`\distancename` and not :math:`\datafidname`.
+        This function directly calls :meth:`deepinv.optim.Distance.prox` for the
+        specific distance function :math:`\distancename`.
+
+        :param torch.Tensor u: Variable :math:`u` at which the gradient is computed.
+        :param torch.Tensor y: Data :math:`y` of the same dimension as :math:`u`.
+        :return: (torch.Tensor) gradient of :math:`d` in :math:`u`, i.e. :math:`\nabla_u\distance{u}{y}`.
+        """
+        out = 0
+        for i, data_fidelity in enumerate(self.data_fidelity_list):
+            out += data_fidelity.prox_d(u, y[i], *args, **kwargs)
+        return out
+
+    def prox_d_conjugate(self, u, y, *args, **kwargs):
+        r"""
+        Computes the proximity operator of the convex conjugate of the distance function :math:`\distance{u}{y}`.
+
+        This function directly calls :meth:`deepinv.optim.Distance.prox_conjugate` for the
+        specific distance function :math:`\distancename`.
+        """
+        out = 0
+        for i, data_fidelity in enumerate(self.data_fidelity_list):
+            out += data_fidelity.prox_d_conjugate(u, y[i], *args, **kwargs)
+        return out
 
 
 class L2(DataFidelity):
@@ -135,7 +253,7 @@ class L2(DataFidelity):
 
     def __init__(self, sigma=1.0):
         super().__init__()
-        self.d = L2Distance()
+        self.d = L2Distance(sigma=sigma)
         self.norm = 1 / (sigma**2)
 
     def prox(self, x, y, physics, *args, gamma=1.0, **kwargs):